# openapi/openapi.json
{
  "components": {
...
    "schemas": {
...
      "Error": {
        "properties": {
          "error": {
            "description": "Stable refusal kind (deny_* / invalid_credentials / …).",
            "type": "string"
          },
          "reason": {
            "description": "Single-line, secret-free human explanation.",
            "type": "string"
          }
        },
        "required": [
          "error",
          "reason"
        ],
        "type": "object"
      },
      "ExportEnvelope": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "description": "Deterministic, sorted section payload.",
            "type": "object"
          },
          "generated_at": {
            "format": "date-time",
            "type": "string"
          },
          "kind": {
            "enum": [
              "profile",
              "members",
              "api_keys",
              "usage",
              "subscription",
              "wallets",
              "audit"
            ],
            "type": "string"
          },
          "organization_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "organization_id",
          "kind",
          "generated_at",
          "data"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "A SaaS session token (login) or a tenant API key. The tenant context is the credential's own organization unless the caller is a platform administrator.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Multi-tenant SaaS contract over the TASK 7A control plane. Trading truth (orders, fills, positions, risk, ledger, HA) remains owned by the TASK 1–6 engines and is exposed read-only to tenants. Billing/usage reads are served through the deterministic exports (`/api/saas/exports?kind=subscription|usage|...`); there are no separate plan-catalogue or usage endpoints.",
    "title": "Sniper Suite SaaS Control Plane",
    "version": "2.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/saas/exports": {
      "get": {
        "operationId": "saas.exportData",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportEnvelope"
                }
              }
            },
            "description": "The export envelope."
          },
          "401": {
            "$ref": "#/components/schemas/Error"
          },
          "403": {
            "$ref": "#/components/schemas/Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Export one deterministic, tenant-scoped section (export.create; audit section additionally needs audit.read).",
        "tags": [
          "exports"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "the control-plane deployment itself",
      "url": "/"
    }
  ],
...
}